fix merge conflicts

+++ SRC/ShineWiFi-ModBus/WebDebug.cpp
@@ -1,4 +1,5 @@
 #include "WebDebug.h"
+#include <stdint.h>
 
 char acWebDebug[1024] = "";
-int u16WebMsgNo = 0;
+uint16_t u16WebMsgNo = 0;
